requirements.txt
algorithms==0.1.3
backcall==0.1.0
certifi==2018.4.16
chardet==3.0.4
click==6.7
colorama==0.3.9
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
dataclasses==0.6
decorator==4.3.0
diff-match-patch==20121119
Django==2.0.7
django-cors-headers==2.4.0
django-crispy-forms==1.7.2
django-filter==2.0.0
django-formtools==2.1
django-guardian==1.4.9
django-import-export==1.0.1
django-mongonaut==0.2.21
django-pure-pagination==0.3.0
django-redis==4.9.0
django-rest-framework==0.1.0
django-rest-framework-mongoengine==3.3.1
django-reversion==3.0.0
django-suit==0.2.26
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
djangorestframework-jwt-refresh-token==0.5
DjangoUeditor==1.8.143
djongo==1.2.29
et-xmlfile==1.0.1
future==0.16.0
httplib2==0.9.2
idna==2.7
ipython==6.4.0
ipython-genutils==0.2.0
iso8601==0.1.12
itypes==1.1.0
jdcal==1.4
jedi==0.12.1
Jinja2==2.10
Markdown==2.6.11
MarkupSafe==1.0
mongoengine==0.15.0
mysqlclient==1.3.12
numpy==1.14.2
odfpy==1.3.6
openpyxl==2.5.4
pandas==0.22.0
parso==0.3.0
pbkdf2==1.3
peewee==3.4.0
pickleshare==0.7.4
Pillow==5.2.0
prompt-toolkit==1.0.15
psycopg2==2.7.4
pygame==1.9.3
Pygments==2.2.0
PyJWT==1.6.4
pymongo==3.6.1
pyserial==3.4
python-dateutil==2.7.2
pytz==2018.5
PyYAML==3.13
redis==2.10.6
requests==2.19.1
schema==0.6.7
simplegeneric==0.8.1
six==1.11.0
SQLAlchemy==1.2.7
sqlparse==0.2.4
tablib==0.12.1
tornado==5.0.2
traitlets==4.3.2
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.23
utils==0.9.0
virtualenv==15.2.0
wcwidth==0.1.7
xadmin==2.0.1
xlrd==1.1.0
xlwt==1.3.0
xmltodict==0.11.0
xmltojson==0.2.0
